move isMatrixDense() function to scalabindings package
